+++ src/domain/services/reviews/create-review.ts
@@ -3,24 +3,10 @@ import { PgIntegrityConstraintViolation } from '@/db/utils/postgres-errors'
 import { UserNotFoundError } from '@/domain/errors/user-not-found'
 import postgres from 'postgres'
 import type { InsertReviewModel } from '../../entities/review'
-import { insertUserActivity } from '@/db/repositories/user-activities'
 
 export async function createReviewService(params: InsertReviewModel) {
   try {
     const [review] = await insertReview(params)
-
-    await insertUserActivity({
-      userId: review.userId,
-      activityType: 'CREATE_REVIEW',
-      entityId: review.id,
-      entityType: 'REVIEW',
-      metadata: {
-        tmdbId: params.tmdbId,
-        mediaType: params.mediaType,
-        seasonNumber: params.seasonNumber,
-        episodeNumber: params.episodeNumber,
-      },
-    })
 
     return { review }
   } catch (error) {