+++ src/domain/services/user-activities/get-user-activities.spec.ts
@@ -1,9 +1,8 @@
 import { describe, expect, it } from 'vitest'
-
 import { makeUser } from '@/test/factories/make-user'
-import { makeRawUserItem } from '@/test/factories/make-user-item'
 import { getUserActivitiesService } from './get-user-activities'
-import { upsertUserItemService } from '../user-items/upsert-user-item'
+import { randomUUID } from 'node:crypto'
+import { createUserActivity } from './create-user-activity'
 
 describe('get user activities', () => {
   it('should be able to get user activities', async () => {
@@ -11,8 +10,12 @@ describe('get user activities', () => {
 
     await Promise.all(
       Array.from({ length: 21 }, async (_, index) => {
-        const userItem = makeRawUserItem({ status: 'WATCHED', userId: user.id })
-        return await upsertUserItemService(userItem)
+        return await createUserActivity({
+          userId: user.id,
+          activityType: 'WATCH_EPISODE',
+          entityId: randomUUID(),
+          entityType: 'REVIEW',
+        })
       })
     )